+++ evmd/ante/evm_handler.go
@@ -1,13 +1,14 @@
 package ante
 
 import (
+	"github.com/cosmos/evm/ante"
 	evmante "github.com/cosmos/evm/ante/evm"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
 // newMonoEVMAnteHandler creates the sdk.AnteHandler implementation for the EVM transactions.
-func newMonoEVMAnteHandler(options HandlerOptions) sdk.AnteHandler {
+func newMonoEVMAnteHandler(options ante.HandlerOptions) sdk.AnteHandler {
 	decorators := []sdk.AnteDecorator{
 		evmante.NewEVMMonoDecorator(
 			options.AccountKeeper,
@@ -15,9 +16,8 @@ func newMonoEVMAnteHandler(options HandlerOptions) sdk.AnteHandler {
 			options.EvmKeeper,
 			options.MaxTxGasWanted,
 		),
+		ante.NewTxListenerDecorator(options.PendingTxListener),
 	}
-	if options.PendingTxListener != nil {
-		decorators = append(decorators, NewTxListenerDecorator(options.PendingTxListener))
-	}
+	
 	return sdk.ChainAnteDecorators(decorators...)
 }